refactor: split mcp_tools_git.go into area-based files
The 1467-line mcp_tools_git.go has been broken up by area. Each new file
holds the Params types, handlers, and formatters for the tools it owns:

- mcp_tools_repo.go    (428 lines): clone, pull, info, branches, switch,
                                    list_repositories, remove_repository
- mcp_tools_files.go   (389 lines): search_files, list_files,
                                    get_file_content, get_readme_files
- mcp_tools_history.go  (88 lines): list_commits, get_commit_diff
- mcp_tools_config.go  (248 lines): session, batch
- mcp_tools_git.go     (124 lines): now only RegisterGitTools, the tool
                                    registration map

Added mcp_errors.go with errorResult/textResult helpers. All handlers
(including mcp_tools_memo.go) now build CallToolResult through these
two helpers instead of inlining the same five-line struct literal at
roughly 50 sites. Behavior is unchanged; this is purely structural.
